more fixes to status bar and terminal emulator. Setting/saving ansi colors works now
