+++ CodeEditModules/Modules/TerminalEmulator/src/Model/TerminalShellType.swift
